requirements.txt
torch
torchvision
gradio
pillow
tqdm
numpy
